+++ src/cpx_io/cpx_system/cpx_ap/parameter.py
@@ -1,0 +1,13 @@
+from dataclasses import dataclass
+
+
+@dataclass
+class Parameter:
+    """Parameter dataclass"""
+
+    parameter_id: int
+    array_size: int
+    data_type: str
+    default_value: int
+    description: str
+    name: str